+++ src/chrome-scripts/classes/tests/Request.test.js
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { Request } from '../';
-import { logger } from '../../modules';
+import { logger } from '../../utils';
 
 vi.mock('../../modules', () => ({
   logger: {
