+++ src/chrome-scripts/content.js
@@ -1,5 +1,5 @@
 import './assets/content.css';
-import * as features from './features';
+import * as features from './features/content';
 
 async function initFeatures() {
   const settings = await getSettings(Object.keys(features));
