+++ src/chrome-scripts/utils/tests/debounce.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { debounce } from '../';
+import { debounce } from '..';
 
 const DELAY = 1000;
