+++ src/chrome-scripts/utils/tests/logger.test.js
